y default, Django displays the str() of each object. But sometimes it’d be more helpful if we could display individual fields. To do that, u
y default, Django displays the str() of each object. But sometimes it’d
be more helpful if we could display individual fields. To do that, use
the list_display admin option, which is a tuple of field names to
display, as columns, on the change list page for the object: