That adds a search box at the top of the change list. When somebody enters search terms, Django will search the question_text field. You can
That adds a search box at the top of the change list. When somebody
enters search terms, Django will search the question_text field. You can
use as many fields as you’d like – although because it uses a LIKE query
behind the scenes, limiting the number of search fields to a reasonable
number will make it easier for your database to do the search.